is mobile function

+++ app/Http/Controllers/AdminController.php
@@ -594,12 +594,36 @@ class AdminController extends Controller
     public function deleteImageHome($id)
     {
         $home_image = HomeImage::find($id);
-        $home_image->delete();
-        $home_images = HomeImage::get();
-        return view('admin.listImageAdds',compact('home_images'));
+        if($home_image)
+        {
+            $home_image->delete();
+            $home_images = HomeImage::get();
+            return view('admin.listImageAdds',compact('home_images'));
+        }
+        else
+        {
+            $home_images = HomeImage::get();
+            return view('admin.listImageAdds',compact('home_images'));
+        }
+       
     }
 
     // END HOME ADS
+
+    // ADD IS MOBILE FUNCTION
+    public function ismobile() 
+    {
+        $is_mobile = Input::get('is_mobile');
+        $id = Input::get('id');
+        $data = [
+                'id' => $id,
+                'is_mobile' => $is_mobile
+            ];
+        $post = Post::find($id);
+        $post->is_mobile = $is_mobile;
+        $post->save();
+        return json_encode($data);
+    }
     
     //POSTS
 	public function posts(Request $request)
